requirements.txt
flask
redis
pyjwt
werkzeug
flask-cors
